Send TraceProcess attributes on resource
OpenTelemetry allows setting attributes on the resource associated with
each span to attach information about the resource creating those spans,
and the spec also details some conventions for naming [1].

So far, this integration has only been sending the `service.name`
attribute as the only value in resource attributes, but for some tracing
solutions, it's important to be able to specify other attributes, such
as the deployment environment (`deployment.environment`).

This moves the attributes associated with the `TraceProcess` from each
individual span to the resource level, where they are more likely to
belong.

[1] https://opentelemetry.io/docs/reference/specification/resource/semantic_conventions/